docs: standardize solved problem documentation

+++ arrays/0001-two-sum/cpp/Solution.cpp
@@ -7,7 +7,7 @@ public:
     vector<int> twoSum(vector<int>& nums, int target) {
         unordered_map<int, int> seen;
 
-        for (int i = 0; i < nums.size(); i++) {
+        for (int i = 0; i < static_cast<int>(nums.size()); i++) {
             int complement = target - nums[i];
             if (seen.count(complement)) {
                 return { seen[complement], i };
